requirements.txt
fastapi
uvicorn
python-multipart
# Updated to support OpenAI class import for DALL-E integration
openai<2.0.0
python-dotenv
requests
jinja2>=3.1.0
vosk
google-generativeai
google-cloud-speech
google-cloud-texttospeech
websockets
pydub
aiohttp
aiofiles